+++ pytket/docs/changelog.rst
@@ -31,6 +31,7 @@ Minor new features:
 * New ``pytket.passes.auto_rebase_pass`` and ``pytket.passes.auto_squash_pass``
   which attempt to construct rebase and squash passess given a target gate set from known
   decompositions.
+* Add ``get_c_register``, ``get_q_register``, ``c_registers`` and ``q_registers`` methods to ``Circuit``.
 * New ``pytket.passes.NaivePlacementPass`` which completes a basic relabelling of all Circuit Qubit
   not labelled as some Architecture Node to any available Architecture Node